@@ -3,19 +3,18 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^5.16.2",
-    "@testing-library/react": "^12.1.3",
-    "@testing-library/react-hooks": "^7.0.2",
-    "@testing-library/user-event": "^13.5.0",
+    "@reduxjs/toolkit": "^1.8.1",
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^13.2.0",
+    "@testing-library/react-hooks": "^8.0.0",
+    "@testing-library/user-event": "^14.2.0",
     "cross-env": "^7.0.3",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-redux": "^7.2.6",
-    "react-router-dom": "^6.2.2",
-    "react-scripts": "5.0.0",
-    "redux": "^4.1.2",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
+    "react-redux": "^8.0.1",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "5.0.1",
     "redux-logger": "^3.0.6",
-    "redux-thunk": "^2.4.1",
     "sass": "^1.51.0",
     "typescript": "^4.6.4",
     "web-vitals": "^2.1.4"
@@ -23,7 +22,9 @@
   "devDependencies": {
     "@types/jest": "^27.5.1",
     "@types/node": "^17.0.33",
-    "@types/react": "^18.0.9"
+    "@types/react": "^18.0.9",
+    "@types/react-redux": "^7.1.24",
+    "npm-upgrade": "^3.1.0"
   },
   "scripts": {
     "start": "node ./bootstap-start.js start",
@@ -31,6 +32,7 @@
     "test": "react-scripts test --watchAll=false",
     "test:watch": "react-scripts test --watchAll",
     "test:coverage": "react-scripts test --watchAll=false --coverage",
+    "upgrade": "npm-upgrade",
     "eject": "react-scripts eject"
   },
   "eslintConfig": {
